+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented here. The format is based
 - Fixed the navigator panel having a slight background tint (#474)
 - The add component dialogs (e.g. `Add ClutchedPathSpring`) now print any addition errors inside
   the dialog (previously, would print the errors to the log, #476)
+- The 3D scene now has basic support for rendering shadows (#10)
 
 ## [0.3.1] - UNRELEASED